% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AvAtt}
\alias{AvAtt}
\title{AvAtt}
\format{
A data frame with 8 rows and 3 variables:
\describe{
  \item{\code{Region}}{integer factor column, London or England}
  \item{\code{Subject}}{integer factor column of subject element}
  \item{\code{Score}}{double Attainment 8 Score} 
}
}
\source{
\url{https://data.london.gov.uk/economic-fairness/equal-opportunities/ks4-achievement/}
}
\usage{
AvAtt
}
\description{
Average Attainment 8 Scores - 2016/17. Average score per pupil in each element
}
\keyword{datasets}
