% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EURef}
\alias{EURef}
\title{EU Referendum Results}
\format{A data frame with 33 rows and 7 variables:
\describe{
  \item{\code{Area_Code}}{integer GSS borough code}
  \item{\code{Area}}{integer Borugh name}
  \item{\code{Leave}}{integer Leave vote}
  \item{\code{Remain}}{integer Remain vote}
  \item{\code{PropLeave}}{double Leave vote as a proportion}
  \item{\code{CatLeave}}{integer Leave vote proportion split into categories}
  \item{\code{geometry}}{list geometry of Borough}  
}}
\source{
\url{https://data.london.gov.uk/dataset/eu-referendum-results}

\url{https://data.london.gov.uk/dataset/statistical-gis-boundary-files-london}
}
\usage{
EURef
}
\description{
Results from the 23rd June 2016 referendum on membership of the 
European Union for each local authority in London
}
\details{
Contains Ordnance Survey data © Crown copyright and database right [2015]
}
\keyword{datasets}
