% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EYFSP}
\alias{EYFSP}
\title{Achievement in EYFSP by Ethnicity}
\format{
A data frame with 12 rows and 3 variables:
\describe{
  \item{\code{Ethnicity}}{integer ethinicity of pupils}
  \item{\code{Region}}{integer London or England}
  \item{\code{eyfsp}}{integer percentage achieving the expected standard} 
}
}
\source{
\url{https://data.london.gov.uk/economic-fairness/equal-opportunities/school-readiness/}
}
\usage{
EYFSP
}
\description{
Percentage achieving at least the expected standard in Equal Opportunities
School Readiness Early Years Foundation Stage tests in all ELGs (2017)
}
\keyword{datasets}
