% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_highlight.R
\docType{data}
\name{GeomGLAAnnotate}
\alias{GeomGLAAnnotate}
\title{GeomGLAAnnotate}
\format{An object of class \code{GeomGLAAnnotate} (inherits from \code{GeomText}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.}
\usage{
GeomGLAAnnotate
}
\description{
Geom to be used for adding annotations to plots.
}
\keyword{datasets}
