% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_line.R
\docType{data}
\name{GeomGLALine}
\alias{GeomGLALine}
\title{GeomGLALine}
\format{
An object of class \code{GeomGLALine} (inherits from \code{GeomLine}, \code{GeomPath}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
GeomGLALine
}
\description{
Geom to be used with ggla_line()
}
\keyword{datasets}
