% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_highlight.R
\docType{data}
\name{GeomGLAPointHighlight}
\alias{GeomGLAPointHighlight}
\title{GeomGLAPointHighlight}
\format{
An object of class \code{GeomGLAPointHighlight} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
GeomGLAPointHighlight
}
\description{
Geom to be used for highlighting points on a line chart with ggla_highlight()
}
\keyword{datasets}
