% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_sf.R
\docType{data}
\name{GeomGLASf}
\alias{GeomGLASf}
\title{GeomGLASf}
\format{An object of class \code{GeomGLASf} (inherits from \code{GeomSf}, \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomGLASf
}
\description{
Geom to be used with ggla_sf()
}
\keyword{datasets}
