% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_highlight.R
\docType{data}
\name{GeomGLATextHighlight}
\alias{GeomGLATextHighlight}
\title{GeomGLATextHighlight}
\format{
An object of class \code{GeomGLATextHighlight} (inherits from \code{GeomText}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
GeomGLATextHighlight
}
\description{
Geom to be used for adding labels to points on a chart with ggla_highlight()
}
\keyword{datasets}
