% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_theme_colours.R
\name{get_theme_colours}
\alias{get_theme_colours}
\title{get_theme_colours()}
\usage{
get_theme_colours()
}
\value{
A named list of colours
}
\description{
Fetches the colours for the current GLA theme
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
