% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_axisat0.R
\name{ggla_axisat0}
\alias{ggla_axisat0}
\title{ggla_axisat0}
\usage{
ggla_axisat0(gla_theme = "default")
}
\arguments{
\item{gla_theme}{Either "default" or "inverse", Default: 'default'}
}
\description{
Forces the y-axis to y=0 and adds a strong axis line
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{geom_abline}}
}
