% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_donut.R
\name{ggla_donut}
\alias{ggla_donut}
\title{ggla_donut}
\usage{
ggla_donut(data = NULL, mapping = NULL, stat = "count",
  title = NULL, ...)
}
\arguments{
\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{stat}{The statistical transformation to use on the data for this layer, 
as a string. Should be either 'count', for the number of cases in each group, or 'identity' to represent values in the data. Default: 'count'}

\item{title}{String title to be used for plot. Will appear in the middle of the donut, Default: NULL}

\item{...}{Other arguments to be passed on to geom_bar()}
}
\description{
This creates a donut chart, similar to a pie chart.
}
\details{
This works by creating a stacked bar chart and converting to polar coordinates.
}
\section{Aesthetics}{

ggla_donut() understands the following aesthetics (required aesthetics in bold).
\itemize{
    \item \strong{y}
    \item label \emph{(will appear outside the donut)}
    \item alpha
    \item colour
    \item fill
    \item group
    \item shape
    \item size
    \item stroke
}
}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{geom_bar}}
}
