% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_highlight.R
\name{ggla_highlight}
\alias{ggla_highlight}
\title{ggla_highlight}
\usage{
ggla_highlight(filter_type = "end", x_filt = NULL, y_filt = NULL,
  data = NULL, mapping = NULL, geom = GeomPointHighlight,
  position = "identity", ...)
}
\arguments{
\item{filter_type}{One of string "end", "start", "max", "min" or "xy", Default: 'end'}

\item{x_filt}{If filter_type = "xy" use this to select which x-values to highlight, Default: NULL}

\item{y_filt}{If filter_type = "xy" use this to select which y-values to highlight, Default: NULL}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{geom}{Override the default connection between geom_highlight() and GeomPointHighlight, Default: GeomPointHighlight}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function., Default: 'identity'}

\item{...}{Other arguments to be passed on to geom_bar()}
}
\description{
Highlight selected data points
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{geom_point}}
}
