% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_horizbar.R
\name{ggla_horizbar}
\alias{ggla_horizbar}
\title{ggla_horizbar}
\usage{
ggla_horizbar(
  ...,
  stat = "identity",
  position = "stack",
  to100 = FALSE,
  gla_theme = "default"
)
}
\arguments{
\item{...}{other parameters to be passed to geom_bar}

\item{stat}{The statistical transformation to use on the data for this layer, as a string, Default: "identity"}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{to100}{boolean, If TRUE a context layer will be added below to bar chart up to 100, Default: FALSE}

\item{gla_theme}{Either "default" or "inverse", Default: 'default'}
}
\description{
Create a horizontal bar chart
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{geom_bar}}
}
