% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_labelline.R
\name{ggla_labelline}
\alias{ggla_labelline}
\title{ggla_labelline}
\usage{
ggla_labelline(
  x1 = NULL,
  x2 = NULL,
  y1 = NULL,
  y2 = NULL,
  gla_theme = "default"
)
}
\arguments{
\item{x1}{starting x-value, not needed for lines perpendicular to x-axis, Default: NULL}

\item{x2}{final x-value, only needed lines which aren't perpendicular or parallel to axis, Default: NULL}

\item{y1}{starting y-value, not needed for lines perpendicular to y-axis, Default: NULL}

\item{y2}{final y-value, only needed lines which aren't perpendicular or parallel to axis, Default: NULL}

\item{gla_theme}{Either "default" or "inverse", Default: 'default'}
}
\description{
Add a dotted line labelled a specific point
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{geom_abline}}
 \code{\link[ggplot2]{geom_segment}}
}
