% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_plotly_config.R
\name{ggla_plotly_config}
\alias{ggla_plotly_config}
\title{ggla_plotly_config}
\usage{
ggla_plotly_config(plot, ...)
}
\arguments{
\item{plot}{A plotly plot, created either using plot_ly or ggplotly.}

\item{...}{Other parameters passed to plotly::config}
}
\value{
A plotly plot
}
\description{
Adds default ggla config to a plotly plot
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
