% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_plotly_settings.R
\name{ggla_plotly_settings}
\alias{ggla_plotly_settings}
\title{ggla_plotly_settings}
\usage{
ggla_plotly_settings(
  plot,
  gla_theme = "default",
  x_axis_title = FALSE,
  y_axis_title = FALSE,
  xgridlines = FALSE,
  legend = TRUE,
  base_size = 14,
  ...
)
}
\arguments{
\item{plot}{A plotly plot, created either using plot_ly or ggplotly.}

\item{gla_theme}{Either "default" or "inverse", Default: 'default'}

\item{x_axis_title}{boolean, If TRUE an x-axis label/title can be added, Default: FALSE}

\item{y_axis_title}{boolean, If TRUE a y-axis label/title can be added, Default: FALSE}

\item{xgridlines}{boolean, If TRUE, theme will include gridlines on the x-axis, Default: FALSE}

\item{legend}{boolean, If TRUE, theme will include a legend at the top left of the plot, Default: TRUE}

\item{base_size}{base font size, Default: 14}

\item{...}{Other parameters passed to plotly::layout}
}
\value{
A plotly plot
}
\description{
AAdds default ggla layout, style and config to a plotly plot
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
