% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_plotly_style.R
\name{ggla_plotly_style}
\alias{ggla_plotly_style}
\title{ggla_plotly_style}
\usage{
ggla_plotly_style(plot, gla_theme = "default", base_size = 14, ...)
}
\arguments{
\item{plot}{A plotly plot, created either using plot_ly or ggplotly.}

\item{gla_theme}{Either "default" or "inverse", Default: 'default'}

\item{base_size}{base font size, Default: 14}

\item{...}{Other parameters passed to plotly::style}
}
\value{
A plotly plot
}
\description{
Adds default ggla style to a plotly plot
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
