% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_sf.R
\name{ggla_sf}
\alias{ggla_sf}
\title{ggla_sf}
\usage{
ggla_sf(gla_theme = "default", ...)
}
\arguments{
\item{gla_theme}{Either "default" or "inverse", Default: 'default'}

\item{...}{paramters to pass to geom_sf}
}
\description{
Wrapper for ggplot2::geom_sf simplyfying coordinates and appearance
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
