% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gla_colours}
\alias{gla_colours}
\title{gla_colours}
\format{named list}
\usage{
gla_colours
}
\description{
list of colours used in GLA palettes
}
\details{
list of hex codes
}
\keyword{datasets}
