% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gla_dark}
\alias{gla_dark}
\title{gla_dark}
\format{named list}
\usage{
gla_dark
}
\description{
list of colours used in the gla_dark theme
}
\details{
list of hex codes of colours used for data and non-data elements of the gla_dark theme
}
\keyword{datasets}
