% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gla_default}
\alias{gla_default}
\title{gla_default}
\format{
named list
}
\usage{
gla_default
}
\description{
list of colours used in the gla_default theme
}
\details{
list of hex codes of colours used for data and non-data elements of the gla_default theme
}
\keyword{datasets}
