% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gla_inverse}
\alias{gla_inverse}
\title{gla_inverse}
\format{
named list
}
\usage{
gla_inverse
}
\description{
list of colours used in the gla_inverse theme
}
\details{
list of hex codes of colours used for data and non-data elements of the gla_inverse theme
}
\keyword{datasets}
