% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gla_light}
\alias{gla_light}
\title{gla_light}
\format{named list}
\usage{
gla_light
}
\description{
list of colours used in the gla_light theme
}
\details{
list of hex codes of colours used for data and non-data elements of the gla_light theme
}
\keyword{datasets}
