% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Internet}
\alias{Internet}
\title{Internet users in London}
\format{A data frame with 2 rows and 3 variables:
\describe{
  \item{\code{locat}}{character London}
  \item{\code{usage}}{integer usage level, used/not used in the last 3 months}
  \item{\code{numUsers}}{double number of users} 
}}
\source{
\url{https://www.ons.gov.uk/businessindustryandtrade/itandinternetindustry/datasets/internetusers}
}
\usage{
Internet
}
\description{
Survery of internet use in the last 3 months
}
\keyword{datasets}
