% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LDNUK}
\alias{LDNUK}
\title{Gender Pay Gap}
\format{An object of class \code{data.frame} with 42 rows and 3 columns.}
\source{
\url{https://data.london.gov.uk/economic-fairness/labour-market/gender-pay-gap/}
}
\usage{
LDNUK
}
\description{
Gender Pay Gap - Total (Median) - London VS UK
}
\keyword{datasets}
