% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_align_labs.R
\name{left_align_labs}
\alias{left_align_labs}
\title{left_align_labs}
\usage{
left_align_labs(plot)
}
\arguments{
\item{plot}{ggplot2 object}
}
\description{
DEPRECATED. The functionality is now included in `theme_gla`.
Left aligns titles, subtitles and captions to the left of the whole plot (default is to left align to just the plot area)
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{is.ggplot}},\code{\link[ggplot2]{ggplotGrob}}
 \code{\link[ggplotify]{as.ggplot}}
}
