% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm_to_pt}
\alias{mm_to_pt}
\title{mm_to_pt}
\format{double}
\usage{
mm_to_pt
}
\description{
Factor converting mm to pt
}
\details{
Various ggplot sizes are set using mms, this factor will adjust this to use pts instead.
}
\keyword{datasets}
