% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_gla_pals.R
\name{show_gla_pals}
\alias{show_gla_pals}
\title{show_gla_pals}
\usage{
show_gla_pals(gla_theme = "default", inc_div = FALSE)
}
\arguments{
\item{gla_theme}{Either "default" or "inverse", Default: "default"}

\item{inc_div}{boolean, If TRUE, output will include diverging colour palettes, Default: FALSE}
}
\description{
Display the available GLA palettes
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
