% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_donut.R
\name{ggla_donut}
\alias{ggla_donut}
\title{ggla_donut}
\usage{
ggla_donut(data = NULL, mapping = NULL, stat = "count",
  title = NULL, gla_theme = "default", base_size = 14, ...)
}
\arguments{
\item{data}{The data to be displayed in this layer. There are three options: If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot(). A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created. A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data, Default: NULL}

\item{mapping}{Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping, Default: NULL}

\item{stat}{The statistical transformation to use on the data for this layer, as a string. Should be either 'count', for the number of cases in each group, or 'identity' to represent values in the data. Default: 'count'}

\item{title}{String title to be used for plot. Will appear in the middle of the donut, Default: NULL}

\item{gla_theme}{Either "default" or "inverse", Default: 'default'}

\item{base_size}{base font size, Default: 14}

\item{...}{Other arguments to be passed on to geom_bar()}
}
\description{
This creates a donut chart, similar to a pie chart.
}
\details{
This works by creating a stacked bar chart and converting to polar coordinates.
}
\section{Aesthetics}{

ggla_donut() understands the following aesthetics (required aesthetics in bold).
\itemize{
    \item \strong{y}
    \item label \emph{(will appear outside the donut)}
    \item alpha
    \item colour
    \item fill
    \item group
    \item shape
    \item size
    \item stroke
}
}

\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{geom_bar}}
}
