% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_highlightarea.R
\name{ggla_highlightarea}
\alias{ggla_highlightarea}
\title{ggla_highlightarea}
\usage{
ggla_highlightarea(
  xmin = -Inf,
  xmax = Inf,
  ymin = -Inf,
  ymax = Inf,
  gla_theme = "default"
)
}
\arguments{
\item{xmin}{lower x-extent of highlighted area, Default: -Inf}

\item{xmax}{higher x-extent of highlighted area, Default: Inf}

\item{ymin}{lower y-extent of highlighted area, Default: -Inf}

\item{ymax}{higher y-extent of highlighted area, Default: Inf}

\item{gla_theme}{Either "default" or "inverse", Default: 'default'}
}
\description{
Add a shaded rectangle behind a plot.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
#' @seealso
 \code{\link[ggplot2]{annotate}}
}
