% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggla_line.R
\name{ggla_line}
\alias{ggla_line}
\title{ggla_line}
\usage{
ggla_line(...)
}
\arguments{
\item{...}{parameters passed to geom_line}
}
\description{
Wrapper for ggplot2::geom_line with round line ends
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{geom_path}}
}
