% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gla_pal.R
\name{gla_pal}
\alias{gla_pal}
\title{gla_pal}
\usage{
gla_pal(gla_theme = NULL, palette_type = "categorical",
  main_colours = "mayoral", n = 6, muted = FALSE, inc0 = FALSE)
}
\arguments{
\item{gla_theme}{One of "light" or "dark". If NULL will try to pick up current theme, or default to light, Default: NULL}

\item{palette_type}{One of "categorical", "quantitative", "highlight" or "diverging", Default: 'categorical'}

\item{main_colours}{One or more of "blue", "pink", "green", "red", "yellow", "orange", "purple" or "mayoral" as a string or list, Default: 'mayoral'}

\item{n}{Number of colours in the palette. If palette_type = "Diverging", this is the number of colours on each side of the diverging scale (so 2n (+1) colours in total will be returned). If palette_type = "Highlight" gla_pal will return main_colours + (n - length(main_colours)) context colours. Default: 6, Default: 6}

\item{muted}{boolean, If TRUE a muted version of the gla colours will be used, Default: FALSE}

\item{inc0}{boolean, If TRUE an additional colour representing zero will be added to a diverging palette. This results in a palette of length 2n + 1, Default: FALSE}
}
\description{
Generates palettes using the GLA colours
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[grDevices]{colorRamp}}
}
