% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gla.R
\name{theme_gla}
\alias{theme_gla}
\title{theme_gla}
\usage{
theme_gla(
  gla_theme = "default",
  x_axis_title = FALSE,
  y_axis_title = FALSE,
  xgridlines = FALSE,
  legend = TRUE,
  base_size = 14
)
}
\arguments{
\item{gla_theme}{Either 'default' or 'inverse', Default: 'default'}

\item{x_axis_title}{boolean, If TRUE an x-axis label/title can be added, Default: FALSE}

\item{y_axis_title}{boolean, If TRUE a y-axis label/title can be added, Default: FALSE}

\item{xgridlines}{boolean, If TRUE, theme will include gridlines on the x-axis, Default: FALSE}

\item{legend}{boolean, If TRUE, theme will include a legend at the top left of the plot, Default: TRUE}

\item{base_size}{base font size, Default: 14}
}
\description{
Controls the non-data display to be consistent with our style guide
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
